pyproject.toml
[tool.poetry]
name = "dl-learning-repo"
version = "0.1.0"
description = "A repo where I try and implement DL concepts I learn."
authors = ["Selvin Selbaraju <selvinoselbaraju@gmail.com>"]
readme = "README.md"
packages = [ {include = "src"} ]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.6.0"
numpy = "^2.2.4"
datasets = "^3.5.1"
timm = "^1.0.15"
matplotlib = "^3.10.1"
lightning = "^2.5.1.post0"
tensorboard = "^2.19.0"
pydantic = "^2.11.4"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
